Add feature to List Group.
Users should be able to list the groups they
have created to view them.

Implementing the list group feature with reference
to List feature that was previously there.

Both functionality works similarly in the sense
they are both listing objects to the users.

No frontend have been done yet and should be
implemented in the future.